moved products js to one folder
