+++ web/src/Components/index.js
@@ -1,8 +1,8 @@
 import Home from "./Home";
 import Navigation from "./Navigation";
 import About from "./About";
-import Product from "./Product";
-import AddProduct from "./Add-Product";
-import DetiledProd from "./Detiled-Product";
+import Product from "./Products/Product";
+import AddProduct from "./Products/Add-Product";
+import DetiledProd from "./Products/Detiled-Product";
 
 export { Home, Navigation, About, Product, AddProduct, DetiledProd };
